+++ SampleApps/SPMTest/SPMTest/ViewController.swift
@@ -9,7 +9,6 @@ import BraintreeThreeDSecure
 import BraintreeUnionPay
 import BraintreeVenmo
 import PayPalDataCollector
-import CardinalMobile
 
 class ViewController: UIViewController {