+++ Sources/BraintreeThreeDSecure/BTProxyCardinal.h
@@ -1,7 +1,13 @@
+#if __has_include(<Braintree/BraintreeThreeDSecure.h>)
+#import <Braintree/BTThreeDSecureV2UICustomization.h>
+#else
+#import <BraintreeThreeDSecure/BTThreeDSecureV2UICustomization.h>
+#endif
+
 NS_ASSUME_NONNULL_BEGIN
 
 // https://flint.tools/blog/finding-a-weak-linking-solution.html
-// To support SPM without a Cardinal.xcframework version, we created proxy protocols for Cardinal classes which are substituted for the actual CardinalMobile classes at runtime.
+// To support SPM without an xcframework version of Cardinal, we created proxy protocols for Cardinal classes which are substituted for the actual CardinalMobile classes at runtime.
 
 typedef NS_ENUM(NSUInteger, BTProxyCardinalSessionEnvironment) {
     BTProxyCardinalSessionEnvironmentStaging,
@@ -25,6 +31,7 @@ typedef NS_ENUM(NSUInteger, BTProxyCardinalResponseActionCode) {
 
 @protocol BTProxyCardinalSessionConfiguration <NSObject>
 @property (nonatomic, assign) BTProxyCardinalSessionEnvironment deploymentEnvironment;
+@property (nonatomic, strong) id uiCustomization;
 @end
 
 typedef void (^BTProxyCardinalSessionSetupDidCompleteHandler)(NSString *consumerSessionId);