+++ Sources/BraintreeThreeDSecure/Public/BraintreeThreeDSecure/BTThreeDSecureRequest.h
@@ -1,9 +1,11 @@
 #if __has_include(<Braintree/BraintreeThreeDSecure.h>)
 #import <Braintree/BTPaymentFlowRequest.h>
 #import <Braintree/BTPaymentFlowDriver.h>
+#import <Braintree/BTThreeDSecureV2UICustomization.h>
 #else
 #import <BraintreePaymentFlow/BTPaymentFlowRequest.h>
 #import <BraintreePaymentFlow/BTPaymentFlowDriver.h>
+#import <BraintreeThreeDSecure/BTThreeDSecureV2UICustomization.h>
 #endif
 
 @class BTThreeDSecureRequest;
@@ -95,10 +97,8 @@ typedef NS_ENUM(NSInteger, BTThreeDSecureVersion){
 
 /**
  Optional. UI Customization for 3DS2 challenge views.
- 
- @see https://cardinaldocs.atlassian.net/wiki/spaces/CMSDK/pages/873234760/iOS+UI+Customization
  */
-@property (nonatomic, nullable, strong) UiCustomization *uiCustomization;
+@property (nonatomic, nullable, strong) BTThreeDSecureV2UICustomization *v2UICustomization;
 
 /**
  Optional. UI Customization for 3DS1 challenge views.